Add answers processing, saving filed answers, evaluating answers. TODO: Refactor evaluating answers, create GET Test Details different responses, add test time limiting, consider on TestResult entity

+++ DataAccess/Entities/Answers/RightSimpleAnswer.cs
@@ -13,7 +13,7 @@ namespace DataAccess.Entities.Answers
         public int Id { get; set; }
 
         [Required]
-        public float Grade { get; set; }
+        public decimal Grade { get; set; }
 
         [Required]
         public int ResponseOptionId { get; set; }
